requirements.txt
Flask==0.10.
CherryPy==3.8.0
docker-py==1.4.0
Flask-Redis==0.1.0
Flask-RESTful==0.3.4
CouchDB==1.0
redis==2.10.3
